Add source for log server and command-line tool (#1)
* Add source for log server and command-line tool

* Allow a partial test name to fetch multiple logs at once

Fetches logs for all names that start with the entered test name.

* Add configuration name to server and command line input. Allow wildcards

This makes the configuration name a mandatory part of the URL fetching a
log, and an optional argument to the command line tool.
Both the test name and configuration name inputs may now end with '*',
indicating they match any name that starts with that prefix.

* Update server.dart

* Address comments on pull request.

Adds LICENCE and other files. Adds correct headers to files.
Improves error handling.