Update results_feed.proto, removing fields and correcting name
The field called "gerrit_change" was actually written to the database
with the name "review" in TryResult objects, and was unused in other
objects.  Renamed to "review".

The "review_path" field was redundant to the review and patchset
fields, so it is removed and its uses replaced by them.

The "trivial_blamelist" field was not used for anything.

"review" field added to a Commit object, to link a commit that lands a
CL back to the Gerrit review of that CL.

Change-Id: Ic7811a5f244a99e6d52772b408187f31b8042c42
Reviewed-on: https://dart-review.googlesource.com/c/dart_ci/+/126881
Reviewed-by: Alexander Thomas <athom@google.com>