Implement keyword subscriptions in the GitHub label notifier
We store per-repository list of keywords in the Firestore - if any of the
keywords are found in the body of the newly opened issue then this is
considered as if the issue was labeled with a specified label.

We plan to use this to get early notifications about dart related issues
filed in the flutter/flutter repository: having words like `gen_snapshot`,
SIGSEGV, etc to notify everybody subscribed to dependency: dart label.

We are not planning to have any UI for editing keyword subscriptions for now.

Change-Id: Ie22342f0a34a46c3165c21f65998439cc5f5d581
Reviewed-on: https://dart-review.googlesource.com/c/dart_ci/+/128861
Reviewed-by: Alexander Thomas <athom@google.com>