[current results ui] Limit vertical size of filter box, make scrollable
Keep the app usable when a large number of filter terms are added
by limiting the vertical extent of the filter box and making it
scrollable. The filter term input field is moved below the filter chips.

A size overflow in the summary numbers at the bottom of the screen is fixed.

Change-Id: Ia06f95f0bf0a343aee5bb847645b5912f04c79bd
Reviewed-on: https://dart-review.googlesource.com/c/dart_ci/+/191540
Reviewed-by: Alexander Thomas <athom@google.com>