[appengine] Add service to resolve source file location from test name
This CL adds a 'dart-ci.appspot.com/test/' URL to redirect to the source
of a test given by name for a given revision.

The test source is computed by matching the test name with the known
structure of tests in the SDK. References to co19 and third party
packages are resolved by looking up the target revision of that
repository in the DEPS file.

The target for SDK tests can be either github (default) or GOB (which
allows to lookup revisions of CL patchsets).

Change-Id: I628ee91bf962151cb66fd6e69ca5559901811d55
Reviewed-on: https://dart-review.googlesource.com/c/dart_ci/+/200181
Commit-Queue: Karl Klose <karlklose@google.com>
Reviewed-by: William Hesse <whesse@google.com>