[current_results_ui] Add link to test source
Adds a link to the source of the test that produced the result.

Additionally, in the case where all results have the same
revision, a source icon is displayed next to the history
icon. Otherwise, only the link next to the configuration name is
shown.

Change-Id: I4c198791fddc22a5a0a801b51024e40a53cda0b6
Reviewed-on: https://dart-review.googlesource.com/c/dart_ci/+/200841
Commit-Queue: Karl Klose <karlklose@google.com>
Reviewed-by: William Hesse <whesse@google.com>