[current results ui] Give scrollable widgets unique controllers
Multiple scrollable widgets were using the same default
ScrollController, which caused exceptions to be thrown
and instability in the application.

Now each widget has its own ScrollController.
The controllers are not saved in a stateful widget
because these scrollable objects are refreshed
with new data when the app redraws.

Because horizontal scrolling of text using the mouse
is disabled on the web, the scrolling of long test
names has been removed. It was not working.

Change-Id: I34e1a701039fa5fb3860166494816a0458bff916
Reviewed-on: https://dart-review.googlesource.com/c/dart_ci/+/213360
Commit-Queue: William Hesse <whesse@google.com>
Reviewed-by: Alexander Thomas <athom@google.com>