Improve android version detection
It is important to detect when crash stack
originates from Android 11 or newer because
newer versions are shipping a fixed unwinding
library. Older versions were producing stack
dumps which required PC adjustment, but
this was fixed in Android 11. If we apply
PC adjustment to a newer stack that
makes it meaningless.

Thus we add a heuristic to detect Android >=11
based on the app installation path.

Change-Id: I65b7d86d4ecc9410c4067c8cbdaa6c81400c599a
Reviewed-on: https://dart-review.googlesource.com/c/dart_ci/+/217200
Reviewed-by: Alexander Thomas <athom@google.com>
Commit-Queue: Slava Egorov <vegorov@google.com>