[builder] Pin SDK version to 2.17.6
Pinning to force the recipe to use this exact versions. Later
have codesigning issues on macOS (https://dartbug.com/49783).

Note that the pinning in pubspec.yaml is better than pinning in
the recipe because a change to pubspec.yaml will trigger new builds
of the package but a change to the recipe will not.

Bug: b/242960194
Change-Id: I109bc64fcdd6ba526e7c2d09e1a87097bc89b34f
Reviewed-on: https://dart-review.googlesource.com/c/dart_ci/+/256120
Commit-Queue: Alexander Thomas <athom@google.com>
Reviewed-by: William Hesse <whesse@google.com>