[current results ui] Replace "text" button with "copy to clipboard"
"text" button currently can't be used, because the text is usually quite
large and rendering it takes forever.

And once rendered, you can't do anything with it other than copying it
to clipboard with the provided button, as browser's "search" can't
search in the rendered text.

This PR replaces the "text" button to "copy to clipboard as text".

In the URL `.../#/filter=dart2wasm`, this copies 142,227 lines of text
in about a second.

I think ideally we would let the user download the text as file, but the
text is generated from in-memory data on demand, so we don't have a URL
for it.

Change-Id: Ib7919a86a8ee511b7215575912c71ab0a31e6c8d
Reviewed-on: https://dart-review.googlesource.com/c/dart_ci/+/424460
Reviewed-by: Alexander Thomas <athom@google.com>